Standardized Case Influence on Parameter Estimates

Gets a lavaan_rerun() output and computes the standardized changes in selected parameters for each case.

Usage

est_change(rerun_out, parameters = NULL)

Arguments

rerun_out

The output from lavaan_rerun().

parameters

A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). Supports specifying an operator to select all parameters with this operators: ~, ~~, =~, and ~1. This vector can contain both parameter names and operators. If NULL, the default, differences on all free parameters will be computed.

Value

A matrix with the number of columns equal to the number of requested parameters, and the number of rows equal to the number of cases. The row names are the case identification values used in lavaan_rerun(). The elements are the standardized difference. Please see Pek and MacCallum (2011), Equation 7.

Details

For each case, est_change() computes the differences in the estimates of selected parameters with and without this case: (estimate with all case) - (estimate without this case). The differences are standardized by dividing the raw differences by their standard errors (Pek & MacCallum, 2011).

If the analysis is not admissible or does not converge when a case is deleted, NAs will be turned for this case on the differences.

Unlike est_change_raw(), est_change() does not support computing the standardized changes of standardized estimates.

Currently it only supports single-group models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202–228. doi:10.1080/00273171.2011.561068

Author

Shu Fai Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)
#> This is lavaan 0.6-12
#> lavaan is FREE software! Please report any bugs.
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
#> lavaan 0.6-12 ended normally after 2 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                            50
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1              -0.159    0.166   -0.954    0.340
#>     iv2               0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1                0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
#> The expected CPU time is 11.85 second(s).
#> Could be faster if ran in parallel.
# Compute the changes in chisq if a case is removed
out <- est_change(fit_rerun)
# Results excluding a case, for the first few cases
head(out)
#>         m1~iv1        m1~iv2        dv~m1      m1~~m1      dv~~dv        gcd
#> 1  0.055155262 -6.413065e-02  0.073609740 -0.04407439  0.06267672 0.01722394
#> 2  0.012022660  3.317249e-03 -0.008275411 -0.09328814 -0.09338458 0.01767099
#> 3  0.001313759  1.480679e-03 -0.023945052 -0.09898698 -0.08413143 0.01745463
#> 4 -0.031965761  3.702747e-03  0.033571795 -0.07952388 -0.07133589 0.01357006
#> 5  0.116999074  2.621956e-02  0.067261907  0.15520013 -0.08842637 0.05271532
#> 6  0.006981059 -9.150724e-06  0.011033699 -0.09666774 -0.06204791 0.01336716
# Note that these are the differences divided by the standard errors
# The rightmost column, `gcd`, contains the
# generalized Cook's distances (Pek & MacCallum, 2011).
head(out[, "gcd", drop = FALSE])
#>          gcd
#> 1 0.01722394
#> 2 0.01767099
#> 3 0.01745463
#> 4 0.01357006
#> 5 0.05271532
#> 6 0.01336716

# Compute the changes for the paths from iv1 and iv2 to m1
out2 <- est_change(fit_rerun, c("m1 ~ iv1", "m1 ~ iv2"))
# Results excluding a case, for the first few cases
head(out2)
#>         m1~iv1        m1~iv2          gcd
#> 1  0.055155262 -6.413065e-02 5.934620e-03
#> 2  0.012022660  3.317249e-03 1.791500e-04
#> 3  0.001313759  1.480679e-03 4.744496e-06
#> 4 -0.031965761  3.702747e-03 1.030137e-03
#> 5  0.116999074  2.621956e-02 1.628486e-02
#> 6  0.006981059 -9.150724e-06 5.082288e-05
# Note that only the changes in the selected parameters are included.
# The generalized Cook's distance is computed only from the selected
# parameter estimates.