Case Influence on Parameter Estimates

Gets a lavaan_rerun() output and computes the changes in selected parameters for each case if included.

Usage

est_change_raw(rerun_out, parameters = NULL, standardized = FALSE)

Arguments

rerun_out

The output from lavaan_rerun().

parameters

A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). Supports specifying an operator to select all parameters with these operators: ~, ~~, =~, and ~1. This vector can contain both parameter names and operators. More details can be found in the help of pars_id(). If omitted or NULL, the default, changes on all free parameters will be computed.

standardized

If TRUE, the changes in the full standardized solution is returned (type = std.all in lavaan::standardizedSolution()). Otherwise, the changes in the unstandardized solution are returned. Default is FALSE.

Value

A matrix with the number of columns equals to the number of requested parameters, and the number of rows equals to the number of cases. The row names are the case identification values used in lavaan_rerun(). The elements are the raw differences.

Details

For each case, est_change_raw() computes the differences in the estimates of selected parameters with and without this case:

(Estimate with all case) - (Estimate without this case).

The change is the raw change, either for the standardized or unstandardized solution. The change is not divided by standard error. This is a measure of the influence of a case on the parameter estimates if it is included.

If the value of a case is positive, including the case increases an estimate.

If the value of a case is negative, including the case decreases an estimate.

If the analysis is not admissible or did not converge when a case is deleted, NAs will be returned for this case on the differences.

Currently it only supports single-group models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202-228. doi:10.1080/00273171.2011.561068

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448.

Examples

library(lavaan)

# A path model

dat <- pa_dat
mod <-
"
m1 ~ a1 * iv1 + a2 * iv2
dv ~ b * m1
a1b := a1 * b
a2b := a2 * b
"
# Fit the model
fit <- lavaan::sem(mod, dat)

# Fit the model n times. Each time with one case is removed.
# For illustration, do this only for four selected cases
fit_rerun <- lavaan_rerun(fit, parallel = FALSE,
                          to_rerun = c(3, 5, 7, 8))
#> The expected CPU time is 0.27 second(s).
#> Could be faster if run in parallel.
# Compute the changes in parameter estimates if a case is included
# vs. if this case is excluded.
# That is, case influence on parameter estimates.
out <- est_change_raw(fit_rerun)
# Results excluding a case
out
#>         m1~iv1       m1~iv2        dv~m1        m1~~m1       dv~~dv
#> 3 -0.004089566 -0.004033540 -0.003161965 -0.0082045913 -0.008465320
#> 5  0.007023707  0.002756245  0.003511133  0.0063451776 -0.012506623
#> 7 -0.012648374  0.007256864  0.006898715 -0.0002583555 -0.007600108
#> 8  0.006160552  0.006730360  0.002966593 -0.0066786194 -0.012636972
#>      a1b:=a1*b    a2b:=a2*b
#> 3 -0.002807807 -0.003748609
#> 5  0.004363007  0.003247601
#> 7 -0.004973074  0.007302370
#> 8  0.003805893  0.005008979
# Note that these are the differences in parameter estimates.

# The parameter estimates from all cases
(coef_all <- coef(fit))
#>     a1     a2      b m1~~m1 dv~~dv 
#>  0.215  0.522  0.517  0.903  1.321 
# The parameter estimates from manually deleting the third case
fit_no_3 <- lavaan::sem(mod, dat[-3, ])
(coef_no_3 <- coef(fit_no_3))
#>     a1     a2      b m1~~m1 dv~~dv 
#>  0.219  0.526  0.520  0.911  1.330 
# The differences
coef_all - coef_no_3
#>     a1     a2      b m1~~m1 dv~~dv 
#> -0.004 -0.004 -0.003 -0.008 -0.008 
# The first row of `est_change_raw(fit_rerun)`
round(out[1, ], 3)
#>    m1~iv1    m1~iv2     dv~m1    m1~~m1    dv~~dv a1b:=a1*b a2b:=a2*b 
#>    -0.004    -0.004    -0.003    -0.008    -0.008    -0.003    -0.004 

# Compute only the changes of the paths from iv1 and iv2 to m1
out2 <- est_change_raw(fit_rerun, c("m1 ~ iv1", "m1 ~ iv2"))
# Results excluding a case
out2
#>         m1~iv1       m1~iv2
#> 3 -0.004089566 -0.004033540
#> 5  0.007023707  0.002756245
#> 7 -0.012648374  0.007256864
#> 8  0.006160552  0.006730360
# Note that only the changes in the selected paths are included.

# Use standardized = TRUE to compare the differences in standardized solution
out2_std <- est_change_raw(fit_rerun,
                           c("m1 ~ iv1", "m1 ~ iv2"),
                           standardized = TRUE)
out2_std
#>         m1~iv1        m1~iv2
#> 3 -0.002382316 -0.0006520026
#> 5  0.004528809 -0.0014871173
#> 7 -0.009968065  0.0055269863
#> 8  0.004814791  0.0061621814
(est_std_all <- parameterEstimates(fit,
                 standardized = TRUE)[1:2, c("lhs", "op", "rhs", "std.all")])
#>   lhs op rhs std.all
#> 1  m1  ~ iv1   0.178
#> 2  m1  ~ iv2   0.459
(est_std_no_1 <- parameterEstimates(fit_no_3,
                 standardized = TRUE)[1:2, c("lhs", "op", "rhs", "std.all")])
#>   lhs op rhs std.all
#> 1  m1  ~ iv1    0.18
#> 2  m1  ~ iv2    0.46
# The differences
est_std_all$std.all - est_std_no_1$std.all
#> [1] -0.0023823159 -0.0006520026
# The first row of `out2_std`
out2_std[1, ]
#>        m1~iv1        m1~iv2 
#> -0.0023823159 -0.0006520026 

# A CFA model

dat <- cfa_dat
mod <-
"
f1 =~  x1 + x2 + x3
f2 =~  x4 + x5 + x6
f1 ~~ f2
"
# Fit the model
fit <- lavaan::cfa(mod, dat)

# Examine four selected cases
fit_rerun <- lavaan_rerun(fit, parallel = FALSE,
                          to_rerun = c(2, 3, 5, 7))
#> The expected CPU time is 0.33 second(s).
#> Could be faster if run in parallel.
# Compute the changes in parameter estimates if a case is included
# vs. if this case is excluded.
# That is, case influence on parameter estimates.
# For free loadings only
out <- est_change_raw(fit_rerun, parameters = "=~")
out
#>         f1=~x2       f1=~x3       f2=~x5       f2=~x6
#> 2  0.097002689  0.120784635  0.120565308 -0.032974404
#> 3 -0.306700558 -0.171683298 -1.016068284 -0.019729433
#> 5 -0.004259315 -0.032446811  0.005350876 -0.009785925
#> 7 -0.004712562 -0.003127709 -0.025473876 -0.004392278
# For standardized loadings only
out_std <- est_change_raw(fit_rerun, parameters = "=~",
                          standardized = TRUE)
out_std
#>          f1=~x1       f1=~x2       f1=~x3       f2=~x4       f2=~x5
#> 2 -0.0382963977  0.032493785  0.029194005 -0.010603775  0.023408705
#> 3  0.1040289435 -0.059072074  0.013663629  0.094725234 -0.103524346
#> 5  0.0095076022 -0.001009356 -0.009363211 -0.002599708  0.002019141
#> 7  0.0003681766 -0.005085922 -0.001441912  0.003520104 -0.001211756
#>         f2=~x6
#> 2 -0.032267866
#> 3  0.095594495
#> 5 -0.004956718
#> 7  0.002613101

# A latent variable model

dat <- sem_dat
mod <-
"
f1 =~  x1 + x2 + x3
f2 =~  x4 + x5 + x6
f3 =~  x7 + x8 + x9
f2 ~   a * f1
f3 ~   b * f2
ab := a * b
"
# Fit the model
fit <- lavaan::sem(mod, dat)

# Examine four selected cases
fit_rerun <- lavaan_rerun(fit, parallel = FALSE,
                          to_rerun = c(2, 3, 5, 7))
#> The expected CPU time is 0.37 second(s).
#> Could be faster if run in parallel.
# Compute the changes in parameter estimates if a case is included
# vs. if this case is excluded.
# That is, case influence on parameter estimates.
# For structural paths only
out <- est_change_raw(fit_rerun, parameters = "~")
out
#>         f2~f1        f3~f2
#> 2  0.01007898 -0.005603283
#> 3 -0.02986991 -0.005846643
#> 5 -0.01554270 -0.001114845
#> 7 -0.01143535  0.006639105
# For standardized paths only
out_std <- est_change_raw(fit_rerun, parameters = "~",
                          standardized = TRUE)
out_std
#>           f2~f1         f3~f2
#> 2  0.0027616467 -0.0005239221
#> 3 -0.0134192997  0.0038105216
#> 5  0.0001256161 -0.0017810573
#> 7  0.0116978785  0.0043196369