Compute case influence

Get a lavaan_rerun() output and compute the changes in selected parameters and fit measures for each case.

influence_stat(
  rerun_out,
  fit_measures = c("chisq", "cfi", "rmsea", "tli"),
  baseline_model = NULL,
  parameters = NULL,
  mahalanobis = TRUE,
  keep_fit = TRUE
)

Arguments

rerun_out
The output from lavaan_rerun().

fit_measures
The argument fit.measures used in lavaan::fitMeasures. Default is c("chisq", "cfi", "rmsea", "tli").

baseline_model
The argument baseline.model used in lavaan::fitMeasures. Default

parameters
A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). If NULL, the default, differences on all free parameters will be computed.

mahalanobis
If TRUE, will call mahalanobis_rerun() to compute the Mahalanobis distance. Default is TRUE.

keep_fit
If TRUE, will keep the original lavaan output using the full sample as an attribute to the output. It can be used by other functions to extract necessary information. Default is TRUE.

Value

A matrix with the number of columns equal to the number of requested statistics, and the number of rows equal to the number of cases. The row names is the case identification values used in lavaan_rerun(). Please refer to the corresponding functions for further details.

Details

For each case, compute the differences in the estimates of selected parameters and fit measures with and without this case. Can also request measures of extremeness (only Mahalanobis distance is available for now).

It is a wrapper of est_change() and fit_measures_change().

Currently only work for single-sample models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202–228. https://doi.org/10.1080/00273171.2011.561068

See also

Author

S. F. Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

#> lavaan 0.6-9 ended normally after 13 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1              -0.159    0.166   -0.954    0.340
#>     iv2               0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1                0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 

# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)

#> The expected CPU time is 2.3 second(s).
#> Could be faster if ran in parallel.

# Get all default influence stats
out <- influence_stat(fit_rerun)
head(out)

#>          chisq cfi rmsea          tli       m1~iv1        m1~iv2       dv~m1
#> 1  0.364966775   0     0 -0.089423524  0.055155120 -6.413054e-02  0.07360975
#> 2  0.002466591   0     0 -0.001853612  0.012022425  3.317406e-03 -0.00827550
#> 3 -0.025206945   0     0  0.004202146  0.001313708  1.480761e-03 -0.02394497
#> 4 -0.035970625   0     0  0.006676175 -0.031966016  3.702928e-03  0.03357168
#> 5  0.110992116   0     0 -0.028824532  0.116998821  2.621972e-02  0.06726178
#> 6  0.086695524   0     0 -0.022536563  0.006980969 -9.042076e-06  0.01103398
#>        m1~~m1      dv~~dv        gcd        md
#> 1 -0.04407445  0.06267652 0.01722389 2.2538385
#> 2 -0.09328821 -0.09338479 0.01767104 0.2927790
#> 3 -0.09898699 -0.08413154 0.01745465 3.0251076
#> 4 -0.07952396 -0.07133612 0.01357011 0.7810530
#> 5  0.15520004 -0.08842656 0.05271526 2.7091742
#> 6 -0.09666780 -0.06204820 0.01336721 0.4390106