semfindr: Influential Cases in Structural Equation Modeling

Sensitivity analysis in structural equation modeling using influence measures and diagnostic plots. Support leave-one-out casewise sensitivity analysis presented by Pek and MacCallum (2011) doi:10.1080/00273171.2011.561068 and approximate casewise influence using scores and casewise likelihood.

See also

Author

Maintainer: Shu Fai Cheung shufai.cheung@gmail.com (ORCID)

Authors:

Mark Hok Chio Lai marklhc@gmail.com (ORCID)