Standardized case influence on parameter estimates

Get a lavaan_rerun() output and compute the standardized changes in selected parameters for each case.

est_change(rerun_out, parameters = NULL)

Arguments

rerun_out
The output from lavaan_rerun().

parameters
A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). If NULL, the default, differences on all free parameters will be computed.

Value

A matrix with the number of columns equal to the number of requested parameters, and the number of rows equal to the number of cases. The row names are the case identification values used in lavaan_rerun(). The elements are the standardized difference. Please see Pek and MacCallum (2011), Equation 7.

Details

For each case, compute the differences in the estimates of selected parameters with and without this case: (estimate with all case) - (estimate without this case). The differences are standardized by dividing the raw differences by their standard errors (Pek & MacCallum, 2011)/

If the analysis is not admissible or did not converge when a case was deleted, NAs will be turned for this case on the differnces.

Unlike est_change_raw(), est_change() does not support computing the standardized changes of standardized estimates.

Currently only support single-sample models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202–228. https://doi.org/10.1080/00273171.2011.561068

Author

S. F. Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)

#> This is lavaan 0.6-9
#> lavaan is FREE software! Please report any bugs.

dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)

#> lavaan 0.6-9 ended normally after 13 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1              -0.159    0.166   -0.954    0.340
#>     iv2               0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1                0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 

# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)

#> The expected CPU time is 13.8 second(s).
#> Could be faster if ran in parallel.

# Compute the changes in chisq if a case is removed
out <- est_change(fit_rerun)
# Results excluding a case, for the first few cases
head(out)

#>         m1~iv1        m1~iv2       dv~m1      m1~~m1      dv~~dv        gcd
#> 1  0.055155120 -6.413054e-02  0.07360975 -0.04407445  0.06267652 0.01722389
#> 2  0.012022425  3.317406e-03 -0.00827550 -0.09328821 -0.09338479 0.01767104
#> 3  0.001313708  1.480761e-03 -0.02394497 -0.09898699 -0.08413154 0.01745465
#> 4 -0.031966016  3.702928e-03  0.03357168 -0.07952396 -0.07133612 0.01357011
#> 5  0.116998821  2.621972e-02  0.06726178  0.15520004 -0.08842656 0.05271526
#> 6  0.006980969 -9.042076e-06  0.01103398 -0.09666780 -0.06204820 0.01336721

# Note that these are the differences divided by the standard errors
# The rightmost column, `gcd`, contains the
# generalized Cook's distances (Pek & MacCallum, 2011).
head(out[, "gcd", drop = FALSE])

#>          gcd
#> 1 0.01722389
#> 2 0.01767104
#> 3 0.01745465
#> 4 0.01357011
#> 5 0.05271526
#> 6 0.01336721

# Compute the changes for the paths from iv1 and iv2 to m1
out2 <- est_change(fit_rerun, c("m1 ~ iv1", "m1 ~ iv2"))
# Results excluding a case, for the first few cases
head(out2)

#>         m1~iv1        m1~iv2          gcd
#> 1  0.055155120 -6.413054e-02 5.934595e-03
#> 2  0.012022425  3.317406e-03 1.791457e-04
#> 3  0.001313708  1.480761e-03 4.744619e-06
#> 4 -0.031966016  3.702928e-03 1.030153e-03
#> 5  0.116998821  2.621972e-02 1.628481e-02
#> 6  0.006980969 -9.042076e-06 5.082190e-05

# Note that only the changes in the selected parameters are included.
# The generalized Cook's distance is computed only from the selected
# parameter estimates.