Case Influence on Parameter Estimates

Gets a lavaan_rerun() output and computes the changes in selected parameters for each case.

Usage

est_change_raw(rerun_out, parameters = NULL, standardized = FALSE)

Arguments

rerun_out

The output from lavaan_rerun().

parameters

A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). Supports specifying an operator to select all parameters with this operators: ~, ~~, =~, and ~1. This vector can contain both parameter names and operators. If NULL, the default, differences on all free parameters will be computed.

standardized

If TRUE, the changes in the full standardized solution is returned (type = std.all in lavaan::standardizedSolution()). Otherwise, the changes in the unstandardized solution is returned. Default is FALSE.

Value

A matrix with the number of columns equals to the number of requested parameters, and the number of rows equals to the number of cases. The row names are the case identification values used in lavaan_rerun(). The elements are the raw differences.

Details

For each case, est_change_raw() computes the differences in the estimates of selected parameters with and without this case: (estimate with all case) - (estimate without this case). The change is the raw change, either for the standardized or unstandardized solution. The change is not divided by standard error.

If the analysis is not admissible or did not converge when a case is deleted, NAs will be returned for this case on the differences.

Currently it only supports single-group models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202–228. doi:10.1080/00273171.2011.561068

Author

Shu Fai Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
#> lavaan 0.6-12 ended normally after 2 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                            50
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1              -0.159    0.166   -0.954    0.340
#>     iv2               0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1                0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
#> The expected CPU time is 2.45 second(s).
#> Could be faster if ran in parallel.
# Compute the changes in parameter estimates if a case is removed
out <- est_change_raw(fit_rerun)
# Results excluding a case, for the first few cases
head(out)
#>          m1~iv1        m1~iv2        dv~m1       m1~~m1      dv~~dv
#> 1  0.0092480903 -1.047900e-02  0.011830109 -0.008096158  0.01779958
#> 2  0.0020230483  5.428092e-04 -0.001349747 -0.017310035 -0.02737256
#> 3  0.0002233261  2.457774e-04 -0.003906910 -0.018389060 -0.02461341
#> 4 -0.0053785110  6.049958e-04  0.005471026 -0.014714311 -0.02081522
#> 5  0.0192294969  4.184807e-03  0.011198261  0.027396310 -0.02589279
#> 6  0.0011748533 -1.497733e-06  0.001793063 -0.017949627 -0.01807066
# Note that these are the differences in parameter estimates.

# The parameter estimates from all cases
(coef_all <- coef(fit))
#> m1~iv1 m1~iv2  dv~m1 m1~~m1 dv~~dv 
#> -0.159  0.525  0.350  0.901  1.423 
# The parameter estimates from manually deleting the first case
fit_no_1 <- lavaan::sem(mod, dat[-1, ])
(coef_no_1 <- coef(fit_no_1))
#> m1~iv1 m1~iv2  dv~m1 m1~~m1 dv~~dv 
#> -0.168  0.536  0.338  0.909  1.406 
# The differences
coef_all - coef_no_1
#> m1~iv1 m1~iv2  dv~m1 m1~~m1 dv~~dv 
#>  0.009 -0.010  0.012 -0.008  0.018 
# The first row of `est_change_raw(fit_rerun)`
round(out[1, ], 3)
#> m1~iv1 m1~iv2  dv~m1 m1~~m1 dv~~dv 
#>  0.009 -0.010  0.012 -0.008  0.018 

# Compute only the changes of the paths from iv1 and iv2 to m1
out2 <- est_change_raw(fit_rerun, c("m1 ~ iv1", "m1 ~ iv2"))
# Results excluding a case, for the first few cases
head(out2)
#>          m1~iv1        m1~iv2
#> 1  0.0092480903 -1.047900e-02
#> 2  0.0020230483  5.428092e-04
#> 3  0.0002233261  2.457774e-04
#> 4 -0.0053785110  6.049958e-04
#> 5  0.0192294969  4.184807e-03
#> 6  0.0011748533 -1.497733e-06
# Note that only the changes in the selected paths are included.

# Use standardized = TRUE to compare the differences in standardized solution
out2_std <- est_change_raw(fit_rerun, 
                          c("m1 ~ iv1", "m1 ~ iv2"),
                          standardized = TRUE)
head(out2_std)
#>          m1~iv1        m1~iv2
#> 1  0.0072910092 -8.065905e-03
#> 2  0.0014862613  3.105220e-04
#> 3 -0.0016488362  7.628936e-03
#> 4 -0.0043088514 -3.605488e-04
#> 5  0.0177947350 -5.627279e-03
#> 6  0.0008102828 -5.167092e-05
(est_std_all <- parameterEstimates(fit,
                 standardized = TRUE)[1:2, c("lhs", "op", "rhs", "std.all")])
#>   lhs op rhs std.all
#> 1  m1  ~ iv1  -0.125
#> 2  m1  ~ iv2   0.425
(est_std_no_1 <- parameterEstimates(fit_no_1,
                 standardized = TRUE)[1:2, c("lhs", "op", "rhs", "std.all")])
#>   lhs op rhs std.all
#> 1  m1  ~ iv1  -0.133
#> 2  m1  ~ iv2   0.433
# The differences
est_std_all$std.all - est_std_no_1$std.all
#> [1]  0.007291009 -0.008065905
# The first row of `out2_std`
out2_std[1, ]
#>       m1~iv1       m1~iv2 
#>  0.007291009 -0.008065905