Case Influence Measures

Gets a lavaan_rerun() output and computes the changes in selected parameters and fit measures for each case.

Usage

influence_stat(
  rerun_out,
  fit_measures = c("chisq", "cfi", "rmsea", "tli"),
  baseline_model = NULL,
  parameters = NULL,
  mahalanobis = TRUE,
  keep_fit = TRUE
)

Arguments

rerun_out

The output from lavaan_rerun().

fit_measures

The argument fit.measures used in lavaan::fitMeasures. Default is c("chisq", "cfi", "rmsea", "tli").

baseline_model

The argument baseline.model used in lavaan::fitMeasures. Default is NULL.

parameters

A character vector to specify the selected parameters. Each parameter is named as in lavaan syntax, e.g., x ~ y or x ~~ y, as appeared in the columns lhs, op, and rhs in the output of lavaan::parameterEstimates(). If NULL, the default, differences on all free parameters will be computed.

mahalanobis

If TRUE, it will call mahalanobis_rerun() to compute the Mahalanobis distance. Default is TRUE.

keep_fit

If TRUE, it will keep the original lavaan output using the full sample as an attribute to the output. It can be used by other functions to extract necessary information. Default is TRUE.

Value

A matrix with the number of columns equals to the number of requested statistics, and the number of rows equals to the number of cases. The row names are the case identification values used in lavaan_rerun(). Please refer to the help pages of est_change() and fit_measures_change() for details.

Details

For each case, influence_stat() computes the differences in the estimates of selected parameters and fit measures with and without this case. Users can request measures of extremeness (only Mahalanobis distance is available for now).

Measures are computed by est_change() and fit_measures_change().

Currently it only works for single-group models.

References

Pek, J., & MacCallum, R. (2011). Sensitivity analysis in structural equation models: Cases and their influence. Multivariate Behavioral Research, 46(2), 202–228. doi:10.1080/00273171.2011.561068

See also

Author

Shu Fai Cheung (shufai.cheung@gmail.com)

Examples

library(lavaan)
dat <- pa_dat
# For illustration only, select only the first 50 cases
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ iv1 + iv2
dv ~ m1
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
#> lavaan 0.6-12 ended normally after 2 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                            50
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1              -0.159    0.166   -0.954    0.340
#>     iv2               0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1                0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 
# Fit the model n times. Each time with one case removed.
fit_rerun <- lavaan_rerun(fit, parallel = FALSE)
#> The expected CPU time is 3.3 second(s).
#> Could be faster if ran in parallel.
# Get all default influence stats
out <- influence_stat(fit_rerun)
head(out)
#>          chisq cfi rmsea          tli       m1~iv1        m1~iv2        dv~m1
#> 1  0.364966775   0     0 -0.089423524  0.055155262 -6.413065e-02  0.073609740
#> 2  0.002466591   0     0 -0.001853612  0.012022660  3.317249e-03 -0.008275411
#> 3 -0.025206945   0     0  0.004202146  0.001313759  1.480679e-03 -0.023945052
#> 4 -0.035970625   0     0  0.006676175 -0.031965761  3.702747e-03  0.033571795
#> 5  0.110992116   0     0 -0.028824532  0.116999074  2.621956e-02  0.067261907
#> 6  0.086695524   0     0 -0.022536563  0.006981059 -9.150724e-06  0.011033699
#>        m1~~m1      dv~~dv        gcd        md
#> 1 -0.04407439  0.06267672 0.01722394 2.2538385
#> 2 -0.09328814 -0.09338458 0.01767099 0.2927790
#> 3 -0.09898698 -0.08413143 0.01745463 3.0251076
#> 4 -0.07952388 -0.07133589 0.01357006 0.7810530
#> 5  0.15520013 -0.08842637 0.05271532 2.7091742
#> 6 -0.09666774 -0.06204791 0.01336716 0.4390106