Mahalanobis Distance On Observed Predictors

Gets a lavaan_rerun() or lavaan::lavaan() output and computes the Mahalanobis distance for each case using only the observed predictors.

Usage

mahalanobis_predictors(
  fit,
  emNorm_arg = list(estimate.worst = FALSE, criterion = 1e-06)
)

Arguments

fit

It can be the output from lavaan, such as lavaan::cfa() and lavaan::sem(), or the output from lavaan_rerun().

emNorm_arg

A list of argument for norm2::emNorm(). Default is list(estimate.worst = FALSE, criterion = 1e-6). Ignored if there is no missing data on the observed predictors.

Value

A one-column matrix (a column vector) of the Mahalanobis distance for each case. The number of rows equals to the number of cases in the data stored in the fit object.

Details

For each case, mahalanobis_predictors() computes the Mahalanobis distance of each case on the observed predictors.

If there are no missing values, stats::mahalanobis() will be used to compute the Mahalanobis distance.

If there are missing values on the observed predictors, the means and variance-covariance matrices will be estimated by maximum likelihood using norm2::emNorm(). The estimates will be passed to modi::MDmiss() to compute the Mahalanobis distance.

Currently this function only supports single-group models.

References

Béguin, C., & Hulliger, B. (2004). Multivariate outlier detection in incomplete survey data: The epidemic algorithm and transformed rank correlations. Journal of the Royal Statistical Society: Series A (Statistics in Society), 167(2), 275-294.

Mahalanobis, P. C. (1936). On the generalized distance in statistics. Proceedings of the National Institute of Science of India, 2, 49-55.

Schafer, J.L. (1997) Analysis of incomplete multivariate data. Chapman & Hall/CRC Press.

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448.

Examples

library(lavaan)
dat <- pa_dat
# For illustration, select only the first 50 cases.
dat <- dat[1:50, ]
# The model
mod <-
"
m1 ~ a1 * iv1 +  a2 * iv2
dv ~ b * m1
a1b := a1 * b
a2b := a2 * b
"
# Fit the model
fit <- lavaan::sem(mod, dat)
summary(fit)
#> lavaan 0.6.14 ended normally after 1 iteration
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         5
#> 
#>   Number of observations                            50
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 1.768
#>   Degrees of freedom                                 2
#>   P-value (Chi-square)                           0.413
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   m1 ~                                                
#>     iv1       (a1)   -0.159    0.166   -0.954    0.340
#>     iv2       (a2)    0.525    0.162    3.241    0.001
#>   dv ~                                                
#>     m1         (b)    0.350    0.161    2.169    0.030
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .m1                0.901    0.180    5.000    0.000
#>    .dv                1.423    0.285    5.000    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     a1b              -0.056    0.064   -0.873    0.382
#>     a2b               0.184    0.102    1.803    0.071
#> 

md_predictors <- mahalanobis_predictors(fit)
md_predictors
#>            md
#> 1  0.51368056
#> 2  0.14923074
#> 3  2.83133321
#> 4  0.25150133
#> 5  0.30454534
#> 6  0.10381491
#> 7  3.08755116
#> 8  2.72181993
#> 9  0.54153019
#> 10 3.14819584
#> 11 1.94474349
#> 12 0.22873982
#> 13 7.17867021
#> 14 0.60357386
#> 15 0.21400365
#> 16 0.70387430
#> 17 1.46418930
#> 18 2.13977511
#> 19 1.17546928
#> 20 4.37779812
#> 21 2.29802338
#> 22 0.22913500
#> 23 0.65135163
#> 24 1.60222431
#> 25 4.90917278
#> 26 2.11488292
#> 27 4.68505480
#> 28 3.10388592
#> 29 0.08442462
#> 30 0.15400825
#> 31 0.07132173
#> 32 4.15651117
#> 33 5.47904841
#> 34 3.43189596
#> 35 1.76891424
#> 36 0.27906678
#> 37 1.25323201
#> 38 0.60312948
#> 39 0.64736786
#> 40 0.42504779
#> 41 0.49463694
#> 42 1.52287738
#> 43 5.11518904
#> 44 1.65448191
#> 45 6.70651179
#> 46 1.17975143
#> 47 2.89669475
#> 48 0.07678826
#> 49 0.42383439
#> 50 6.29749467