Convert Parameter Syntax to Position or Row Numbers in the Parameter Vector or Table

Converts a vector of lavaan syntax to the ids of parameters in the vector of free parameters or the row numbers in the parameter table.

Usage

pars_id(pars, fit, where = c("coef", "partable"), free_only = TRUE)

Arguments

pars

A character vector of parameters specified in lavaan syntax, e.g., "y ~ x" and f1 =~ x3. Can mix specific parameters with operators. See the Details section.

fit

A lavaan-class object. This object is used to determine the number of groups and the parameters in the model. Only parameters in pars that appear in this model will be considered.

where

Where the values are to be found. Can be "partable" (parameter table) or "coef" (coefficient vector). Default is "coef".

free_only

Wether only free parameters will be kept. Default is TRUE.

Value

A numeric vector of the ids in the column "free" in the parameter table of the fit object.

Details

It supports the following ways to specify the parameters to be included.

lavaan syntax

For example, "y ~ x" denotes the regression coefficient regression y on x. It uses lavaan::lavaanify() to parse the syntax strings.

Operator

For example, "~" denotes all regression coefficients.

It is used by function such as est_change() and fit_measures_change().

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

dat <- sem_dat

library(lavaan)
sem_model <-
"
f1 =~  x1 + x2 + x3
f2 =~  x4 + x5 + x6
f3 =~  x7 + x8 + x9
f2 ~   f1
f3 ~   f2
"

fit_ng <- sem(sem_model, dat)

pars <- c("f1 =~ x2", "f2 =~ x5", "f2 ~ f1")
tmp <- pars_id(pars, fit = fit_ng)
coef(fit_ng)[tmp]
#>    f1=~x2    f2=~x5     f2~f1 
#> 0.5895183 0.7299536 1.1145114 
tmp <- pars_id(pars, fit = fit_ng, where = "partable")
parameterTable(fit_ng)[tmp, ]
#>    id lhs op rhs user block group free ustart exo label plabel start   est
#> 2   2  f1 =~  x2    1     1     1    1     NA   0         .p2. 0.684 0.590
#> 5   5  f2 =~  x5    1     1     1    3     NA   0         .p5. 0.978 0.730
#> 10 10  f2  ~  f1    1     1     1    7     NA   0        .p10. 0.000 1.115
#>       se
#> 2  0.145
#> 5  0.099
#> 10 0.233